app/requirements.txt
streamlit
pymongo
pandas
wordcloud
matplotlib
Pillow
requests
plotly
tabulate